Foreign cities

+++ JeffShared/WeatherClasses.cs
@@ -30,7 +30,7 @@ namespace JeffShared
 	public class Wind
 	{
 		public double speed { get; set; }
-		public int deg { get; set; }
+		public double deg { get; set; }
 	}
 
 	public class Clouds
